Unmanaged - Etag and Custom Read
Entity tag (ETag) field used in optimistic concurrency (paralell / at the same time work) control, particularly when multiple users are working on the same data simultaneously. The ETag field uniquely represents the current state of a specific resource, such as an entity instance. Whenever a change is made to the resource, the ETag field is updated. During a change request, the ETag value sent by the RAP BO consumer is compared with the current value of the ETag field. If the values match, the change is allowed; otherwise, the request is rejected. This mechanism helps prevent data inconsistencies and unintended changes when multiple users are interacting with the same resource.

More Info : https://help.sap.com/doc/abapdocu_cp_index_htm/CLOUD/en-US/abenbdl_etag.htm